Accept non-Haskell QuasiQuotes
Rather than throwing an error (thus making it impossible to use shm in a
function that has a QQ somewhere in it), instead just gracefully fail.

Ideally, the contents of the QuasiQuote would be one context contained
within the overall QQ, but I couldn't figure out how to do that.

Closes #116.