This was trickier than I thought
The calls to move the data constructor up or down should work on record and non-record data declarations. It will swap the appropriate choices depending on where the point is. For example:

data IndexSettingSummary = IndexSettingSummary { sSummaryIndexName :: IndexName
                                               , sSummararyFixedSettings :: IndexSettings
                                               , sSummaryUpdateable :: [UpdatableIndexSetting]
                                               }
                         | Whatever { somethingDumber :: Char
                                    , somethingStupid :: Int
                                    }

With a data declaration, the records, e.g. somethingDumber will only be swapped the a record in the same enclosing squiggly brackets provided the point is on the appropriate node. If you want to swap IndexSettingSummary and Whatever, it will work as expected. These additions also take into account comments and moves them as you would expect.